Add method to change status of checkbox and update it in local storage
